Move export wallet actions from actions.ts
